+++ src/components/BlogNav.vue
@@ -46,7 +46,7 @@
           Hi, I'm Arlex Mart&iacute;nez! <strong>@lexmartinez</strong><br/> Front-end Developer at Batteries911. JavaScript Surfer. NodeJS lover. Focused on web technologies.
           <br/>Author this blog. Guitar enthusiast 🤘
         </p>
-        <p style="text-align: center">Find me on <a class="fa fa-github-alt" href="https://github.com/lexmartinez" target="_blank"></a> &nbsp;
+        <p style="text-align: center">Find me on &nbsp;<a class="fa fa-github-alt" href="https://github.com/lexmartinez" target="_blank"></a> &nbsp;
           <a class="fa fa-twitter" href="https://twitter.com/lex0316" target="_blank"></a> &nbsp;
           <a class="fa fa-linkedin-square" href="https://www.linkedin.com/in/arlex-martinez-6256a4153/" target="_blank"></a>
           , DMs are open!</p>